Rename internal test package
