+++ circuit-foundation/src/commonJvmTest/kotlin/com/slack/circuit/foundation/CircuitContentTest.kt
@@ -16,13 +16,13 @@ import androidx.compose.ui.test.assertTextEquals
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
+import com.slack.circuit.internal.test.Parcelize
+import com.slack.circuit.internal.test.TestContentTags.TAG_COUNT
+import com.slack.circuit.internal.test.TestContentTags.TAG_INCREASE_COUNT
 import com.slack.circuit.runtime.CircuitUiState
 import com.slack.circuit.runtime.presenter.Presenter
 import com.slack.circuit.runtime.screen.Screen
 import com.slack.circuit.runtime.ui.ui
-import com.slack.circuit.test.Parcelize
-import com.slack.circuit.test.TestContentTags.TAG_COUNT
-import com.slack.circuit.test.TestContentTags.TAG_INCREASE_COUNT
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
