+++ circuit-foundation/src/commonJvmTest/kotlin/com/slack/circuit/foundation/NavigableCircuitSaveableStateTest.kt
@@ -7,14 +7,14 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
 import com.slack.circuit.backstack.rememberSaveableBackStack
-import com.slack.circuit.test.TestContentTags.TAG_COUNT
-import com.slack.circuit.test.TestContentTags.TAG_GO_NEXT
-import com.slack.circuit.test.TestContentTags.TAG_INCREASE_COUNT
-import com.slack.circuit.test.TestContentTags.TAG_LABEL
-import com.slack.circuit.test.TestContentTags.TAG_POP
-import com.slack.circuit.test.TestCountPresenter
-import com.slack.circuit.test.TestScreen
-import com.slack.circuit.test.createTestCircuit
+import com.slack.circuit.internal.test.TestContentTags.TAG_COUNT
+import com.slack.circuit.internal.test.TestContentTags.TAG_GO_NEXT
+import com.slack.circuit.internal.test.TestContentTags.TAG_INCREASE_COUNT
+import com.slack.circuit.internal.test.TestContentTags.TAG_LABEL
+import com.slack.circuit.internal.test.TestContentTags.TAG_POP
+import com.slack.circuit.internal.test.TestCountPresenter
+import com.slack.circuit.internal.test.TestScreen
+import com.slack.circuit.internal.test.createTestCircuit
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
