+++ circuit-foundation/src/commonJvmTest/kotlin/com/slack/circuit/foundation/NavigatorTest.kt
@@ -4,8 +4,8 @@ package com.slack.circuit.foundation
 
 import com.google.common.truth.Truth.assertThat
 import com.slack.circuit.backstack.SaveableBackStack
+import com.slack.circuit.internal.test.Parcelize
 import com.slack.circuit.runtime.screen.Screen
-import com.slack.circuit.test.Parcelize
 import kotlin.test.assertFailsWith
 import kotlin.test.fail
 import org.junit.Test
