+++ internal-test-utils/src/androidMain/kotlin/com/slack/circuit/internal/test/Parcelize.android.kt
@@ -1,5 +1,5 @@
 // Copyright (C) 2023 Slack Technologies, LLC
 // SPDX-License-Identifier: Apache-2.0
-package com.slack.circuit.test
+package com.slack.circuit.internal.test
 
 actual typealias Parcelize = kotlinx.parcelize.Parcelize
