+++ internal-test-utils/src/commonMain/kotlin/com/slack/circuit/internal/test/Parcelize.kt
@@ -1,6 +1,6 @@
 // Copyright (C) 2023 Slack Technologies, LLC
 // SPDX-License-Identifier: Apache-2.0
-package com.slack.circuit.test
+package com.slack.circuit.internal.test
 
 @Target(AnnotationTarget.CLASS)
 @Retention(AnnotationRetention.BINARY)
