+++ internal-test-utils/src/commonMain/kotlin/com/slack/circuit/internal/test/TestContent.kt
@@ -1,6 +1,6 @@
 // Copyright (C) 2023 Slack Technologies, LLC
 // SPDX-License-Identifier: Apache-2.0
-package com.slack.circuit.test
+package com.slack.circuit.internal.test
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
