+++ internal-test-utils/src/iosMain/kotlin/com/slack/circuit/internal/test/Parcelize.kt
@@ -1,6 +1,6 @@
 // Copyright (C) 2023 Slack Technologies, LLC
 // SPDX-License-Identifier: Apache-2.0
-package com.slack.circuit.test
+package com.slack.circuit.internal.test
 
 /** A no-op [Parcelize] annotation for non-Android */
 @Target(AnnotationTarget.CLASS)
